requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
drf_spectacular==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-debug-toolbar==4.4.2
flake8
Pillow>=8.2.0
